refactor(tests): consolidate E2E test matrix — 12 files → 10, zero duplicates
Major test restructuring organized by USE CASE:
- test_01_platform_health.py (7)
- test_02_a2a_connectivity.py (8) — NEW
- test_03_credential_security.py (18)
- test_04_sandbox_lifecycle.py (9)
- test_05_multiturn_conversation.py (15) — NEW
- test_06_conversation_resume.py (5) — NEW
- test_07_skill_execution.py (32) — consolidated
- test_08_supervisor_enforcement.py (12)
- test_09_hitl_policy.py (3) — NEW
- test_10_workspace_persistence.py (8) — NEW

New: HITL policy blocking, PVC conversation resume, canonical test data.
Deleted: 7 merged files, zero duplicates remain.
