+++ kagenti/tests/e2e/openshell/test_01_platform_health.py
