+++ kagenti/tests/e2e/openshell/test_03_credential_security.py
@@ -171,48 +171,3 @@ class TestPolicyConfigMapMounted:
         assert "filesystem_policy:" in content or "network_policies:" in content, (
             "Policy missing filesystem or network policy section"
         )
-
-
-class TestSupervisorIntegration:
-    """Tests for when the OpenShell supervisor is the container entrypoint.
-
-    These verify placeholder tokens, supervisor process tree, and
-    credential resolution. Currently the supervisor is not yet deployed
-    as the entrypoint — these tests document the target state.
-    """
-
-    @pytest.mark.parametrize("agent", _AGENTS)
-    def test_supervisor_entrypoint_or_document_gap(self, agent, agent_namespace):
-        """Check if supervisor is PID 1. If not, document the current state."""
-        pod_name = _get_pod_name(agent, agent_namespace)
-        result = _kubectl_exec(pod_name, agent_namespace, ["cat", "/proc/1/cmdline"])
-        if result.returncode != 0:
-            pytest.skip(f"Cannot read /proc/1/cmdline: {result.stderr}")
-
-        cmdline = result.stdout.replace("\x00", " ").strip()
-        if "openshell" in cmdline.lower() or "supervisor" in cmdline.lower():
-            pass  # Supervisor is entrypoint — test passes
-        else:
-            pytest.skip(
-                f"Supervisor not yet entrypoint for {agent}. "
-                f"Current PID 1: {cmdline[:100]}. "
-                f"This will pass once supervisor integration is deployed."
-            )
-
-    @pytest.mark.parametrize("agent", _LLM_AGENTS)
-    def test_placeholder_tokens_or_document_gap(self, agent, agent_namespace):
-        """Check if env vars use openshell:resolve:env: placeholders."""
-        pod_name = _get_pod_name(agent, agent_namespace)
-        result = _kubectl_exec(pod_name, agent_namespace, ["env"])
-        if result.returncode != 0:
-            pytest.skip(f"Cannot exec into {agent}: {result.stderr}")
-
-        env_lines = result.stdout.strip().splitlines()
-        has_placeholder = any("openshell:resolve:env:" in line for line in env_lines)
-        if has_placeholder:
-            pass  # Placeholders present — test passes
-        else:
-            pytest.skip(
-                f"Placeholder tokens not yet in {agent} env. "
-                f"This will pass once supervisor integration is deployed."
-            )
