+++ kagenti/tests/e2e/openshell/test_04_sandbox_lifecycle.py
@@ -283,123 +283,4 @@ class TestSandboxStatusObservability:
         assert len(result.stdout) > 0, "Gateway logs are empty"
 
 
-@pytest.mark.asyncio
-class TestAgentServicePersistence:
-    """A2A equivalent of the proposal's session reconnect validation criterion.
-
-    The proposal requires: "Sandbox survives CLI disconnect and reconnect
-    (if --keep is used)."
-
-    In our A2A-first model, agents are long-running Deployment-backed
-    services, not ephemeral CLI sessions. Persistence means the agent
-    pod remains available and responsive across multiple independent
-    client connections — equivalent to disconnect + reconnect in the
-    CLI model.
-
-    These tests send multiple sequential A2A requests using independent
-    HTTP connections to verify agent availability and response consistency.
-    """
-
-    async def test_agent_responds_across_connections(self, weather_agent_url):
-        """Agent responds correctly across multiple independent connections.
-
-        Each request uses a fresh httpx.AsyncClient (new TCP connection),
-        simulating CLI disconnect + reconnect.
-        """
-        responses = []
-        for i in range(3):
-            async with httpx.AsyncClient() as client:
-                resp = await a2a_send(
-                    client,
-                    weather_agent_url,
-                    "What is the weather in Paris?",
-                    request_id=f"persist-{i}",
-                )
-            assert "result" in resp, f"Request {i}: missing result"
-            text = extract_a2a_text(resp)
-            assert text and len(text) > 5, f"Request {i}: empty response"
-            responses.append(text)
-
-        assert len(responses) == 3, "Not all 3 requests succeeded"
-
-    async def test_agent_stable_after_delay(self, weather_agent_url):
-        """Agent remains responsive after a short idle period.
-
-        Simulates a user disconnecting, waiting, then reconnecting.
-        """
-        import asyncio
-
-        async with httpx.AsyncClient() as client:
-            resp1 = await a2a_send(
-                client,
-                weather_agent_url,
-                "Weather in Berlin?",
-                request_id="before-delay",
-            )
-        assert "result" in resp1
-
-        await asyncio.sleep(5)
-
-        async with httpx.AsyncClient() as client:
-            resp2 = await a2a_send(
-                client,
-                weather_agent_url,
-                "Weather in Madrid?",
-                request_id="after-delay",
-            )
-        assert "result" in resp2
-        text = extract_a2a_text(resp2)
-        assert text and len(text) > 5, "Agent unresponsive after idle period"
-
-    async def test_agent_pod_not_restarted_during_requests(
-        self,
-        weather_agent_url,
-        agent_namespace,
-    ):
-        """Agent pod does not restart between requests — true persistence."""
-        pods_before = kubectl_get_pods_json(agent_namespace)
-        weather_before = [
-            p
-            for p in pods_before
-            if p["metadata"]["name"].startswith("weather-agent")
-            and "-supervised" not in p["metadata"]["name"]
-            and "-build" not in p["metadata"]["name"]
-            and p["status"].get("phase") == "Running"
-        ]
-        if not weather_before:
-            pytest.skip("No running weather-agent pod found")
-
-        restart_count_before = sum(
-            cs.get("restartCount", 0)
-            for p in weather_before
-            for cs in p["status"].get("containerStatuses", [])
-        )
-
-        async with httpx.AsyncClient() as client:
-            await a2a_send(
-                client,
-                weather_agent_url,
-                "Weather in Rome?",
-                request_id="restart-check",
-            )
-
-        pods_after = kubectl_get_pods_json(agent_namespace)
-        weather_after = [
-            p
-            for p in pods_after
-            if p["metadata"]["name"].startswith("weather-agent")
-            and "-supervised" not in p["metadata"]["name"]
-            and "-build" not in p["metadata"]["name"]
-            and p["status"].get("phase") == "Running"
-        ]
-
-        restart_count_after = sum(
-            cs.get("restartCount", 0)
-            for p in weather_after
-            for cs in p["status"].get("containerStatuses", [])
-        )
-
-        assert restart_count_after == restart_count_before, (
-            f"Agent pod restarted during request: "
-            f"restarts before={restart_count_before}, after={restart_count_after}"
-        )
+# TestAgentServicePersistence moved to test_05_multiturn_conversation.py
