+++ kagenti/tests/e2e/openshell/test_08_supervisor_enforcement.py
@@ -228,90 +228,4 @@ class TestOPAPolicyEnforcement:
         assert "ephemeral CA generated" in logs
 
 
-class TestRealGitHubPRReview:
-    """Test agent reviewing a real GitHub PR diff.
-
-    Fetches a real PR diff from GitHub and sends it to the agent
-    for review. Requires OPENSHELL_LLM_AVAILABLE=true and a
-    GitHub token (from GITHUB_TOKEN env or .env.kagenti).
-    """
-
-    @pytest.mark.asyncio
-    @pytest.mark.skipif(
-        os.getenv("OPENSHELL_LLM_AVAILABLE", "").lower() != "true",
-        reason="LLM not available",
-    )
-    async def test_review_real_github_pr(self, claude_sdk_agent_url):
-        """Fetch a real PR diff from kagenti repo and review it."""
-        gh_token = os.getenv("GITHUB_TOKEN", "")
-
-        # Fetch a recent PR diff from kagenti repo
-        headers = {"Accept": "application/vnd.github.v3.diff"}
-        if gh_token:
-            headers["Authorization"] = f"token {gh_token}"
-
-        async with httpx.AsyncClient() as client:
-            # Use a known small PR for predictable testing
-            diff_resp = await client.get(
-                "https://api.github.com/repos/kagenti/kagenti/pulls/1300",
-                headers={**headers, "Accept": "application/vnd.github.v3.diff"},
-                timeout=15.0,
-            )
-
-        if diff_resp.status_code != 200:
-            pytest.skip(f"Cannot fetch PR diff: HTTP {diff_resp.status_code}")
-
-        diff_text = diff_resp.text[:2000]  # First 2KB to keep prompt short
-
-        async with httpx.AsyncClient() as client:
-            resp = await a2a_send(
-                client,
-                claude_sdk_agent_url,
-                f"Review this pull request diff for security and code quality:\n\n```diff\n{diff_text}\n```",
-                request_id="github-pr-review",
-                timeout=120.0,
-            )
-
-        assert "result" in resp
-        text = extract_a2a_text(resp)
-        assert text, "Empty review response"
-        assert len(text) > 50, f"Review too short: {text[:100]}"
-
-    @pytest.mark.asyncio
-    @pytest.mark.skipif(
-        os.getenv("OPENSHELL_LLM_AVAILABLE", "").lower() != "true",
-        reason="LLM not available",
-    )
-    async def test_rca_style_log_analysis(self, claude_sdk_agent_url):
-        """Send CI-style error logs and ask agent for root cause analysis."""
-        fake_ci_logs = """
-[2026-04-22T10:00:00Z] Starting E2E test suite...
-[2026-04-22T10:00:05Z] test_platform_health PASSED
-[2026-04-22T10:00:10Z] test_agent_deploy PASSED
-[2026-04-22T10:00:15Z] test_weather_agent FAILED
-[2026-04-22T10:00:15Z] ERROR: Connection refused: weather-agent.team1.svc:8080
-[2026-04-22T10:00:15Z] TRACEBACK: httpx.ConnectError: All connection attempts failed
-[2026-04-22T10:00:16Z] test_adk_agent FAILED
-[2026-04-22T10:00:16Z] ERROR: Connection refused: adk-agent.team1.svc:8080
-[2026-04-22T10:00:17Z] Pod weather-agent-abc123: CrashLoopBackOff (3 restarts)
-[2026-04-22T10:00:17Z] Pod adk-agent-def456: CreateContainerConfigError
-[2026-04-22T10:00:17Z] Events: secret "litellm-virtual-keys" not found
-"""
-        async with httpx.AsyncClient() as client:
-            resp = await a2a_send(
-                client,
-                claude_sdk_agent_url,
-                f"Analyze these CI logs and identify the root cause of the failures:\n\n```\n{fake_ci_logs}\n```",
-                request_id="rca-logs",
-                timeout=120.0,
-            )
-
-        assert "result" in resp
-        text = extract_a2a_text(resp)
-        assert text, "Empty RCA response"
-        text_lower = text.lower()
-        has_rca = any(
-            kw in text_lower
-            for kw in ["secret", "litellm", "missing", "not found", "root cause", "fix"]
-        )
-        assert has_rca, f"Response doesn't identify root cause: {text[:200]}"
+# TestRealGitHubPRReview moved to test_07_skill_execution.py::TestRealWorldSkillExecution
